be/refactor: pageNumber 추가, BigDecimal -> Double 수정, 들여쓰기 수정

+++ backend/src/main/java/com/techeer/backend/api/feedback/dto/FeedbackCreateRequest.java
@@ -6,8 +6,6 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @AllArgsConstructor
 // @ToString
@@ -18,20 +16,21 @@ public class FeedbackCreateRequest {
 
 	@NotNull(message = "x 좌표는 필수입니다.")
 	@JsonProperty("x_coordinate")
-	private BigDecimal xCoordinate;
+	private Double xCoordinate;
 
 	@NotNull(message = "y 좌표는 필수입니다.")
 	@JsonProperty("y_coordinate")
-	private BigDecimal yCoordinate;
+	private Double yCoordinate;
 
 	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	private BigDecimal xcoordinate = null;
+	private Double xcoordinate = null;
 
 	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	private BigDecimal ycoordinate = null;
+	private Double ycoordinate = null;
 
 	@NotNull(message = "page 번호는 필수입니다.")
 	private int pageNumber;
+
 }
